requirements_tests.txt
vertica-python==1.0.0
PyYAML==5.3.1
docker==4.4.1
